[PVR] Fix PVR inter component dependencies: timer must not depend on timers.
