+++ xbmc/pvr/PVRContextMenus.cpp
@@ -23,6 +23,7 @@
 #include "pvr/recordings/PVRRecordings.h"
 #include "pvr/recordings/PVRRecordingsPath.h"
 #include "pvr/timers/PVRTimers.h"
+#include "pvr/timers/PVRTimersPath.h"
 
 namespace PVR
 {
