+++ xbmc/pvr/timers/CMakeLists.txt
@@ -1,9 +1,11 @@
 set(SOURCES PVRTimerInfoTag.cpp
             PVRTimers.cpp
+            PVRTimersPath.cpp
             PVRTimerType.cpp)
 
 set(HEADERS PVRTimerInfoTag.h
             PVRTimers.h
+            PVRTimersPath.h
             PVRTimerType.h)
 
 core_add_library(pvr_timers)
