+++ xbmc/pvr/windows/GUIWindowPVRTimersBase.cpp
@@ -22,7 +22,8 @@
 #include "pvr/PVRGUIActions.h"
 #include "pvr/PVRManager.h"
 #include "pvr/addons/PVRClients.h"
-#include "pvr/timers/PVRTimers.h"
+#include "pvr/timers/PVRTimerInfoTag.h"
+#include "pvr/timers/PVRTimersPath.h"
 
 using namespace PVR;
 using namespace KODI::MESSAGING;
